requirements-test.txt
attrs==19.3.0
aws-sam-translator==1.19.0
aws-xray-sdk==2.4.3
boto==2.49.0
boto3==1.10.35
botocore==1.13.35
certifi==2019.11.28
cffi==1.13.2
cfn-lint==0.26.0
chardet==3.0.4
cryptography==2.8
docker==4.1.0
docutils==0.15.2
ecdsa==0.14.1
future==0.18.2
idna==2.8
jinja2==2.10.3
jmespath==0.9.4
jsondiff==1.1.2
jsonpatch==1.24
jsonpickle==1.2
jsonpointer==2.0
jsonschema==3.2.0
markupsafe==1.1.1
mock==3.0.5
more-itertools==8.0.2
moto==1.3.14
packaging==19.2
pluggy==0.13.1
py==1.8.0
pyasn1==0.4.8
pycparser==2.19
pyparsing==2.4.5
pyrsistent==0.15.6
pytest==5.3.1
python-dateutil==2.8.0
python-jose==3.0.1
pytz==2019.3
pyyaml==5.2
requests==2.22.0
requests-mock==1.7.0
responses==0.10.7
rsa==4.0
s3transfer==0.2.1
six==1.13.0
sshpubkeys==3.1.0
urllib3==1.25.7
wcwidth==0.1.7
websocket-client==0.56.0
werkzeug==0.16.0
wrapt==1.11.2
xmltodict==0.12.0
